Split DocHarness out into testz-extras
