+++ extras/src/main/scala/testz/extras/DocHarness.scala
@@ -29,19 +29,35 @@
  */
 
 package testz
+package extras
 
-object StdlibSuite {
-  def tests[T](harness: Harness[T]): T = {
-    import harness._
-    section("assert")(
-      test("success") { () =>
-        if (assert(true) eq Succeed) Succeed()
-        else Fail.noMessage
-      },
-      test("failure") { () =>
-        if (assert(false) eq Fail.noMessage) Succeed()
-        else Fail.noMessage
-      }
-    )
+import scala.collection.mutable.ListBuffer
+
+object DocHarness extends PureHarness[λ[R => (String, ListBuffer[String]) => Unit]] {
+  def test[R]
+    (name: String)
+    (assertions: R => Result)
+    : (String, ListBuffer[String]) => Unit = {
+      (indent, buf) =>
+        buf += (indent + "  " + name)
+    }
+
+  // Can't think of a way to document this, so we don't.
+  def allocate[R, I]
+    (init: () => I
+    )(tests: (String, ListBuffer[String]) => Unit
+    ): (String, ListBuffer[String]) => Unit =
+      tests
+
+  def section[R](name: String)(
+    test1: (String, ListBuffer[String]) => Unit,
+    tests: (String, ListBuffer[String]) => Unit*
+  ): (String, ListBuffer[String]) => Unit = {
+    (indent, buf) =>
+      val newIndent = indent + "  "
+      buf += (newIndent + "[" + name + "]")
+      test1(newIndent, buf)
+      tests.foreach(_(newIndent, buf))
   }
+
 }
