+++ tests/src/main/scala/testz/ExtrasSuite.scala
@@ -30,48 +30,31 @@
 
 package testz
 
-import scalaz._, Scalaz._
-import z._, property._
+import extras._
 
-final class PropertySuite {
-  // TODO: perhaps there's a better way to test these than math.
-  val testData = Unfold[Id, Int]((1 to 6): _*).flatMap {
-    i =>
-      val n = i * 10
-      Unfold((n to (n + 5)): _*)
-  }
-
+object ExtrasSuite {
   def tests[T](harness: Harness[T]): T = {
     import harness._
-    section("int ranges")(
-      test("exhaustiveS") { () =>
-        val actualErrors = exhaustiveS[Id, Int](1, 2, 3, 4, 5, 6)(i =>
-          if (i =/= 3) Fail.string(s"$i is not 3")
-          else Succeed()
+    section("document harness")(
+      test("entire harness") { () =>
+        val buf = new scala.collection.mutable.ListBuffer[String]()
+        DocHarness.section("outer section")(
+          DocHarness.section("first inner section")(
+            DocHarness.test[Unit]("first test inside of first inner section")(_ => Succeed),
+            DocHarness.test[Unit]("second test inside of first inner section")(_ => Succeed)
+          ),
+          DocHarness.section("second inner section")(
+            DocHarness.test[Unit]("first test inside of second inner section")(_ => Succeed),
+          )
+        )("  ", buf)
+        assert(buf.result() == List(
+          "    [outer section]",
+          "      [first inner section]",
+          "        first test inside of first inner section",
+          "        second test inside of first inner section",
+          "      [second inner section]",
+          "        first test inside of second inner section")
         )
-        val expectedErrors = Fail.strings(List(
-          "1 is not 3",
-          "2 is not 3",
-          "4 is not 3",
-          "5 is not 3",
-          "6 is not 3"
-        ))
-        assert(actualErrors === expectedErrors)
-      },
-      test("exhaustiveU") { () =>
-        val actualErrors = exhaustiveU[Id, Int](testData) { i =>
-          if (i % 5 === 1) Fail.string(s"$i is 1 modulo 5")
-          else Succeed()
-        }
-        val expectedErrors = Fail.strings(List(
-          "61 is 1 modulo 5",
-          "51 is 1 modulo 5",
-          "41 is 1 modulo 5",
-          "31 is 1 modulo 5",
-          "21 is 1 modulo 5",
-          "11 is 1 modulo 5"
-        ))
-        assert(actualErrors === expectedErrors)
       }
     )
   }
