+++ tests/src/main/scala/testz/PropertySuite.scala
@@ -30,17 +30,48 @@
 
 package testz
 
-object StdlibSuite {
+import scalaz._, Scalaz._
+import z._, property._
+
+object PropertySuite {
+  // TODO: perhaps there's a better way to test these than math.
+  val testData = Unfold[Id, Int]((1 to 6): _*).flatMap {
+    i =>
+      val n = i * 10
+      Unfold((n to (n + 5)): _*)
+  }
+
   def tests[T](harness: Harness[T]): T = {
     import harness._
-    section("assert")(
-      test("success") { () =>
-        if (assert(true) eq Succeed) Succeed()
-        else Fail.noMessage
+    section("int ranges")(
+      test("exhaustiveS") { () =>
+        val actualErrors = exhaustiveS[Id, Int](1, 2, 3, 4, 5, 6)(i =>
+          if (i =/= 3) Fail.string(s"$i is not 3")
+          else Succeed()
+        )
+        val expectedErrors = Fail.strings(List(
+          "1 is not 3",
+          "2 is not 3",
+          "4 is not 3",
+          "5 is not 3",
+          "6 is not 3"
+        ))
+        assert(actualErrors === expectedErrors)
       },
-      test("failure") { () =>
-        if (assert(false) eq Fail.noMessage) Succeed()
-        else Fail.noMessage
+      test("exhaustiveU") { () =>
+        val actualErrors = exhaustiveU[Id, Int](testData) { i =>
+          if (i % 5 === 1) Fail.string(s"$i is 1 modulo 5")
+          else Succeed()
+        }
+        val expectedErrors = Fail.strings(List(
+          "61 is 1 modulo 5",
+          "51 is 1 modulo 5",
+          "41 is 1 modulo 5",
+          "31 is 1 modulo 5",
+          "21 is 1 modulo 5",
+          "11 is 1 modulo 5"
+        ))
+        assert(actualErrors === expectedErrors)
       }
     )
   }
